requirements.txt
pyyaml
ua-parser
user-agents
django-user-agents
